+++ secstr/secure_string.go
@@ -45,11 +45,15 @@ func NewSecureString(data interface{}) (*String, error) {
 
 // IsEmpty returns false if string is empty
 func (s *String) IsEmpty() bool {
-	return len(s.Data) == 0
+	return s == nil || len(s.Data) == 0
 }
 
 // Destroy destroys data
 func (s *String) Destroy() error {
+	if s == nil {
+		return nil
+	}
+
 	return destroySecureString(s)
 }