+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 ## Changelog
 
+### 12.43.0
+
+* `[terminal]` Added method `ReadPasswordSecure` for reading password into a secure string
+* `[secstr]` Fixed panic in `IsEmpty` and `Destroy` if string struct is nil
+
+### 12.42.1
+
+* Module path set to `v12`
+
 ### 12.42.0
 
 * `[fsutil]` Improved helpers `CopyFile`, `MoveFile` and `CopyDir`